requirements.txt
dlib>=19.24.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0
argparse
pathlib
